Updated Application To Rename vs Delete
Changed the program so it will rename the id.dat file instead of deleting it. This allows you to retrieve the file later and get data out of it in case you want to see the info about the account previously associated with the card (This isn't everything about the account, of course).